+++ ui/src/alfresco/DockerContainerList.tsx
@@ -17,11 +17,7 @@ import { viewContainer, openAlfrescoInBrowser } from '../helper/cli';
 import { resources } from '../helper/resources';
 
 import React from 'react';
-import {
-  AlfrescoStates,
-  ServiceStore,
-  ContainerState,
-} from './alfrescoServices';
+import { AlfrescoStates, ServiceStore, ContainerState } from './types';
 
 const startAlfresco = async () => {
   await openAlfrescoInBrowser();
